% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLE.R
\name{checkRLE}
\alias{checkRLE}
\title{plot relative log expression normalization}
\usage{
checkRLE(raw_counts, analysis)
}
\arguments{
\item{raw_counts}{matrix of raw count data}

\item{analysis}{object with config options saved from the template format}
}
\value{
2 ggplots
}
\description{
Returns plots of count frequencies before and after normalization
}
\examples{
  checkRLE(raw_counts, analysis)

}
