% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSeurat2Pb.R
\name{convertSeurat2Pb}
\alias{convertSeurat2Pb}
\title{Create \code{bulk} list from \code{SeuratObject}}
\usage{
convertSeurat2Pb(
  seurat_obj,
  cluster = NULL,
  sample = NULL,
  arrange_by = NULL,
  sample_md = NULL,
  design_str = "~ 0",
  split_clusters = FALSE
)
}
\arguments{
\item{seurat_obj}{a Seurat or SeuratObject class data object. Must have counts in its RNA assay}

\item{cluster}{character string specifying the column of \code{object@meta.data} that contains single cell cluster labels to split on.}

\item{sample}{character string specifying the column of \code{object@meta.data} that contains sample information of the single cell data.}

\item{arrange_by}{character string specifying the column of \code{object@meta.data} that contains a factor or character vector to arrange the metadata data table by. Note that counts columns will be in same order as samples table to enable cleaner visualizations.}

\item{sample_md}{data.frame (or tibble) with all necessary factors for the design and downstream model fitting}

\item{design_str}{model design string. Default is "~ 0".}

\item{split_clusters}{whether to isolate clusters in separate objects so that you can fit models and define contrasts on each separately. See \code{Details} section.}
}
\description{
Creates a \code{bulk} list compatible with \code{gencoreBulk} functions supporting the \code{sc_template} pseudobulk workflow.
}
\details{
If \code{split_clusters = FALSE} (default), a \code{bulk} list is create with all clusters included. This is helpful if you intend to draw contrasts between clusters. In practice, this most helpful when few clusters are considered and the study design is relatively simple (few factors with few levels), as large numbers of clusters yield unwieldy sample tables and contrast matrices.

If \code{split_clusters = TRUE}, a list of \code{bulk} lists is create with each cluster isolated in its own \code{bulk} list. This allows you to fit the model on each cluster separately, but make it impossible to draw contrasts between clusters. This can be helpful as an exploratory step when there are many clusters of uncertain importance and/or for complex designs with many factor levels.
}
\examples{
\dontrun{
# Create sample metadata table
sample_md <- s.cd8@meta.data \%>\%
  dplyr::distinct(animalID, group, challenge) \%>\%
  as_tibble()

# Create a pseudobulk object with all clusters kept together
bulk <- createPBobj(s.cd8, cluster = "s.fullwnn1.0", 
                    sample = "animalID", sample_md = sample_md, 
                    arrange_by = "challenge", split_clusters = F,
                    design_str = "~ 0")
                    
# Create a pseudobulk object with all clusters kept together
bulk <- createPBobj(s.cd8, cluster = "s.fullwnn1.0", 
                    sample = "animalID", sample_md = sample_md, 
                    arrange_by = "challenge", split_clusters = T,
                    design_str = "~ 0")
}
}
