% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{createResTable}
\alias{createResTable}
\title{Create results table from \code{bulk} list object}
\usage{
createResTable(fit, contr.matrix)
}
\arguments{
\item{fit}{list containing a linear model fit produced by \code{lmFit}, \code{lm.series}, \code{gls.series} or \code{mrlm}.
  For \code{topTable}, \code{fit} should be an object of class \code{MArrayLM} as produced by \code{lmFit} and \code{eBayes}.}

\item{contr.matrix}{Contrast matrix created by \code{limma::makeContrasts()}}
}
\description{
Create tibble of top genes from linear model fit for each contrast \code{\link[limma:topTable]{limma::topTable()}}.
}
\examples{
\dontrun{
bulk <- runVoomLmFit(bulk, contr.matrix=contr.matrix, 
                     sample.weights = TRUE, 
                     block = bulk$dge$samples$SubjectID)
resultsTable <- createResTable(bulk, contr.matrix)
}

}
