% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFilteredDESeqResults.R
\name{.maxMinFilter}
\alias{.maxMinFilter}
\title{Filter genes with low read counts from results calculations}
\usage{
.maxMinFilter(dds, group_samples, min_counts = 0, max_min_samples = 1)
}
\arguments{
\item{dds}{A DESeq object}

\item{group_samples}{List of character vectors containing sample names used
in comparison groups}

\item{min_counts}{Minimum number of counts for a sample to not be ignored}

\item{max_min_samples}{Maximum number of 'zero' samples allowed in a group
to still perform the calculation}
}
\value{
A vector of filtering criteria to pass to \code{DESeq2::results()}
}
\description{
Remove genes with insufficient samples for calculating results. If a
group of samples has more than \code{max_min_samples} with fewer than
\code{min_counts} reads for a gene, the gene will be filtered out.
}
