% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFilteredDESeqResults.R
\name{extractFilteredDESeqResults}
\alias{extractFilteredDESeqResults}
\title{Extract results from processed DESeq2 object using Greg's filtering}
\usage{
extractFilteredDESeqResults(
  comparison,
  intgroup,
  filt_groups,
  test = "Wald",
  dds = analysis$dds,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{comparison}{Value to pass to [DESeq2::results()] as either a `contrast`
or a `name`}

\item{intgroup}{Metadata column on which to draw groups for filtering}

\item{filt_groups}{Character vector of groups to draw from `intgroup`}

\item{test}{Test to use, default is `Wald`}

\item{dds}{DESeq2 object to draw results from}

\item{alpha}{The significance cutoff used for optimizing the independent 
filtering (by default 0.05). If the adjusted p-value cutoff (FDR) will be a 
value other than 0.05, alpha should be set to that value.}

\item{...}{Additional arguments passed to [DESeq2::results()]}
}
\value{
object of class DESeqResults
}
\description{
Calls DESeq2::results with the filter specified using Greg's
minMaxFilter function. Greg's filter dynamically filters results for 
minimum supporting reads based on the samples involved in the result calculation.
This function is intended to reduce the
overall code used to extract all results of interest.
}
