% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitVoomLm}
\alias{fitVoomLm}
\title{Running model fitting procedure using voom + LmFit}
\usage{
fitVoomLm(
  bulkObj,
  contr.matrix,
  block = NULL,
  sample.weights = TRUE,
  var.design = NULL,
  var.group = NULL,
  plotVoom = TRUE
)
}
\arguments{
\item{bulkObj}{List object with raw counts in \code{bulkObj$dge$counts} and design matrix in \code{bulkObj$md$design}.}

\item{contr.matrix}{Contrast matrix creating by \code{limma::makeContrasts()}}

\item{block}{
    vector or factor specifying a blocking variable on the samples.
    Has length equal to the number of \code{ncol(counts)}.
    Samples within each block are assumed to be correlated.}

\item{sample.weights}{
    logical value, if \code{TRUE} then empirical sample quality weights will be estimated.}

\item{var.design}{
    optional design matrix for the sample weights.
    Defaults to the sample-specific model whereby each sample has a distinct variance.}

\item{var.group}{
    optional vector or factor indicating groups to have different array weights.
    This is another way to specify \code{var.design} for groupwise sample weights.}

\item{plotVoom}{logical, should a plot of the mean-variance trend be displayed?}
}
\description{
Model fitting and contrast extraction using voom + LmFit procedure as implemented in \code{\link[edgeR:voomLmFit]{edgeR::voomLmFit()}}.
}
\details{
This is a convenience wrapper around \code{\link[edgeR:voomLmFit]{edgeR::voomLmFit()}}, \code{\link[limma:contrasts.fit]{limma::contrasts.fit()}}, \code{\link[limma:eBayes]{limma::eBayes()}}, and \code{\link[limma:decideTests]{limma::decideTests()}}.
}
\section{Explanation of workflow:}{
\subsection{edgeR::voomLmFit() (Adapted from edgeR::voomLmFit() documentation):}{

This function adapts the limma voom method (Law et al, 2014) to allow for loss of residual degrees of freedom due to exact zero counts (Lun and Smyth, 2017).

The function is analogous to calling \code{voom} followed by \code{duplicateCorrelation} and \code{lmFit} except for the modified residual df values and residual standard deviation sigma values.

If block is specified, then the intra-block correlation is estimated using \code{duplicateCorrelation}. In that case, the voom weights and the intra-block correlation are each estimated twice to achieve effective convergence.

Empirical sample quality weights will be estimated if \code{sample.weights=TRUE} or if \code{var.design} or \code{var.group} are non-NULL (Liu et al 2015). In that case, \code{voomLmFit} is analogous to running \code{voomWithQualityWeights} followed by \code{lmFit}.
}

\subsection{limma::contrasts.fit():}{

Re-orient the fitted model object from the coefficients of the original design matrix to the set of contrasts defined above in contr.matrix
}

\subsection{limma::eBayes():}{

Run empirical Bayes moderation; borrowing information across all the genes to obtain more precise estimates of gene-wise variability
}

\subsection{plotSA():}{

Plot the model's residual variances against average expression values; demonstrating that the variance is no longer dependednt on the mean expression level
}

\subsection{decideTests():}{

Identify which genes are significantly differentially expressed for each contrast from a fit object containing p-values and test statistics
}
}

\examples{
\dontrun{
## Fit a model with repeated measures design with SubjectID treated as a random effect
bulk <- runVoomLmFit(bulk, 
                          contr.matrix=contr.matrix, 
                          sample.weights = TRUE, 
                          block = bulk$dge$samples$SubjectID)
}

}
