% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencoreBulk-package.R
\docType{package}
\name{gencoreBulk-package}
\alias{gencoreBulk}
\alias{gencoreBulk-package}
\title{gencoreBulk: NPRC Genomics Core Bulk RNA Analysis Utilities}
\description{
A shared codebase for custom R functions we find useful for bulk RNA analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yerkes-gencore/gencoreBulk}
}

}
\keyword{internal}
