% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSampleTable.R
\name{generateSampleTable}
\alias{generateSampleTable}
\title{Generate minimal sampletable for use in QC template}
\usage{
generateSampleTable(
  config,
  fileID_pattern = "([^/])/.+",
  sampleID_pattern = "^p\\\\d+.(s\\\\d+).+"
)
}
\arguments{
\item{config}{QC config with alignmentDir and STARreadSuffix specified}

\item{fileID_pattern}{Pattern to pull for FileID}

\item{sampleID_pattern}{Pattern to pull for SampleID}
}
\value{
A dataframe with columns FileID and SampleID
}
\description{
A convenience function to start a sample table with the minimum
requirements to run the QC template. You may need to manually adjust the
output to add additional metadata or trim sampleIDs.
}
\examples{
\dontrun{
sampletable <- generateSampleTable(analysis$qc_config)
}
}
