% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVolcanoPlot.R
\name{generateVolcanoPlot}
\alias{generateVolcanoPlot}
\title{Generate volcano plot from DESeq2 results object}
\usage{
generateVolcanoPlot(
  result,
  labels = 20,
  FCcutoff = log2(1.3),
  pCutoff = 0.05,
  title = NULL,
  caption = NULL,
  subtitle = NULL,
  ...
)
}
\arguments{
\item{result}{DESeqResults object}

\item{labels}{If integer, the top N genes (by adjusted p value) to label.
If a character vector, labels those genes regardless of values.}

\item{FCcutoff}{Cut-off for absolute log2 fold-change. Vertical lines will
be drawn at the negative and positive values of log2FCcutoff.}

\item{pCutoff}{Cut-off for statistical significance. A horizontal line
will be drawn at -log10(pCutoff).}

\item{title}{Plot title.}

\item{caption}{Plot caption.}

\item{subtitle}{Plot subtitle.}

\item{...}{
  Arguments passed on to \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano}}
  \describe{
    \item{\code{toptable}}{A data-frame of test statistics (if not, a data frame,
an attempt will be made to convert it to one). Requires at least
the following: column for variable names (can be rownames); a column
for log2 fold changes; a column for nominal or adjusted p-value.}
    \item{\code{lab}}{A column name in \code{toptable} containing variable names.
Can be \code{rownames(toptable)}.}
    \item{\code{x}}{A column name in \code{toptable} containing log2 fold changes.}
    \item{\code{y}}{A column name in \code{toptable} containing nominal or adjusted
p-values.}
    \item{\code{selectLab}}{A vector containing a subset of lab.}
    \item{\code{xlim}}{Limits of the x-axis.}
    \item{\code{ylim}}{Limits of the y-axis.}
    \item{\code{xlab}}{Label for x-axis.}
    \item{\code{ylab}}{Label for y-axis.}
    \item{\code{axisLabSize}}{Size of x- and y-axis labels.}
    \item{\code{titleLabSize}}{Size of plot title.}
    \item{\code{subtitleLabSize}}{Size of plot subtitle.}
    \item{\code{captionLabSize}}{Size of plot caption.}
    \item{\code{pCutoffCol}}{Column name of statistical significance values to be used as
the cut-off. A typical usage situation would be to pass nominal [un-adjusted]
p-values as 'y', but adjusted p-values as pCutoffCol. In this way, a
plot is generated via -log10(unadjusted p-value), but cut-offs based on
adjusted p-values.}
    \item{\code{cutoffLineType}}{Line type for \code{FCcutoff} and \code{pCutoff}
('blank', 'solid', 'dashed', 'dotted', 'dotdash', 'longdash', 'twodash').}
    \item{\code{cutoffLineCol}}{Line colour for \code{FCcutoff} and \code{pCutoff}.}
    \item{\code{cutoffLineWidth}}{Line width for \code{FCcutoff} and \code{pCutoff}.}
    \item{\code{pointSize}}{Size of plotted points for each variable. Can be
a single value or a vector of sizes.}
    \item{\code{labSize}}{Size of labels for each variable.}
    \item{\code{labCol}}{Colour of labels for each variable.}
    \item{\code{labFace}}{Font face of labels for each variable.}
    \item{\code{boxedLabels}}{Logical, indicating whether or not to draw labels in
boxes.}
    \item{\code{parseLabels}}{Logical, indicating whether or not to parse expressions
in labels}
    \item{\code{shape}}{Shape of the plotted points. Either a single value for
all points, or 4 values corresponding to the default 4 legend labels
specified by \code{legendLabels}.}
    \item{\code{shapeCustom}}{Named vector / key-value pairs that will over-ride the
default shape scheme. The order must match that of \code{toptable}.
Names / keys relate to groups / categories; values relate to shape encodings.}
    \item{\code{col}}{Colour shading for plotted points, corresponding to
the default 4 legend labels specified by \code{legendLabels}.}
    \item{\code{colCustom}}{Named vector / key-value pairs that will over-ride the
default colour scheme. The order must match that of \code{toptable}.
Names / keys relate to groups / categories; values relate to colour.}
    \item{\code{colAlpha}}{Alpha for purposes of controlling colour transparency of
variable points.}
    \item{\code{colGradient}}{If activated, over-rides the default discrete colour scheme
and replaces it with a continous scheme that shades based on nominal or 
adjusted p-value specified by \code{y}. For example, c('red2', 'blue2').}
    \item{\code{colGradientBreaks}}{Break-points for the two colours specified by
colGradient.}
    \item{\code{colGradientLabels}}{Labels for the break-points specified by
colGradientBreaks.}
    \item{\code{colGradientLimits}}{Limits of the colour scheme specified by
colGradient, i.e., max and min possible p-values.}
    \item{\code{legendLabels}}{Plot legend text labels.}
    \item{\code{legendPosition}}{Position of legend ('top', 'bottom', 'left',
'right').}
    \item{\code{legendLabSize}}{Size of plot legend text.}
    \item{\code{legendIconSize}}{Size of plot legend icons / symbols.}
    \item{\code{legendDropLevels}}{Logical, drop unused factor levels from legend.}
    \item{\code{encircle}}{A vector of variable names to encircle. Requires installation
of package \code{\link[ggalt:geom_encircle]{ggalt}}.}
    \item{\code{encircleCol}}{Colour of the encircled line.}
    \item{\code{encircleFill}}{Colour fill of the encircled region.}
    \item{\code{encircleAlpha}}{Alpha for purposes of controlling colour transparency of
encircled region.}
    \item{\code{encircleSize}}{Line width of the encircled line.}
    \item{\code{shade}}{A vector of variable names to shade.}
    \item{\code{shadeFill}}{Colour of shaded regions.}
    \item{\code{shadeAlpha}}{Alpha for purposes of controlling colour transparency of
shaded region.}
    \item{\code{shadeSize}}{Size of the shade contour lines.}
    \item{\code{shadeBins}}{Number of bins for the density of the shade.}
    \item{\code{drawConnectors}}{Logical, indicating whether or not to connect plot
labels to their corresponding points by line connectors.}
    \item{\code{widthConnectors}}{Line width of connectors.}
    \item{\code{typeConnectors}}{Have the arrow head open ('open') or filled ('closed')?}
    \item{\code{endsConnectors}}{Which end of connectors to draw arrow head? ('last',
'first', 'both').}
    \item{\code{lengthConnectors}}{Length (size) of the connector arrowheads.}
    \item{\code{colConnectors}}{Line colour of connectors and line segments.}
    \item{\code{max.overlaps}}{Equivalent of max.overlaps in ggrepel. Set to
'Inf' to always display all labels when drawConnectors = TRUE.}
    \item{\code{maxoverlapsConnectors}}{See max.overlaps.}
    \item{\code{min.segment.length}}{When drawConnectors = TRUE, specifies the minimum
length of the connector line segments.}
    \item{\code{directionConnectors}}{direction in which to draw connectors.
'both', 'x', or 'y'.}
    \item{\code{arrowheads}}{Logical, indicating whether or not to draw arrow heads or
or just have straight lines.}
    \item{\code{hline}}{Draw one or more horizontal lines passing through this/these
values on y-axis. For single values, only a single numerical value is
necessary. For multiple lines, pass these as a vector, e.g., c(60,90).}
    \item{\code{hlineType}}{Line type for \code{hline} ('blank', 'solid', 'dashed', 'dotted',
'dotdash', 'longdash', 'twodash').}
    \item{\code{hlineCol}}{Colour of \code{hline}.}
    \item{\code{hlineWidth}}{Width of \code{hline}.}
    \item{\code{vline}}{Draw one or more vertical lines passing through this/these
values on x-axis. For single values, only a single numerical value is
necessary. For multiple lines, pass these as a vector, e.g., c(60,90).}
    \item{\code{vlineType}}{Line type for \code{vline} ('blank', 'solid', 'dashed', 'dotted',
'dotdash', 'longdash', 'twodash').}
    \item{\code{vlineCol}}{Colour of \code{vline}.}
    \item{\code{vlineWidth}}{Width of \code{vline}.}
    \item{\code{gridlines.major}}{Logical, indicating whether or not to draw major
gridlines.}
    \item{\code{gridlines.minor}}{Logical, indicating whether or not to draw minor
gridlines.}
    \item{\code{border}}{Add a border for just the x and y axes ('partial') or the
entire plot grid ('full')?}
    \item{\code{borderWidth}}{Width of the border on the x and y axes.}
    \item{\code{borderColour}}{Colour of the border on the x and y axes.}
    \item{\code{raster}}{Logical, indicating whether to rasterize the geom_point layer. 
Requires installation of \code{\link[ggrastr:geom_point_rast]{ggrastr}}.}
  }}
}
\value{
A \code{\link{ggplot2}} object.
}
\description{
Generate volcano plot from DESeq2 results object. Labels can be custom set,
or specified as the top N genes by adjusted p value.
}
