% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopNGenes.R
\name{getTopNGenes}
\alias{getTopNGenes}
\title{Title}
\usage{
getTopNGenes(
  result,
  N = 30,
  min_logFC = log2(1.5),
  min_padj = analysis$analysis_config$alpha
)
}
\arguments{
\item{result}{DESeqResults object}

\item{N}{Number of genes to return}

\item{min_logFC}{Minimum absolute value of log2FoldChange}

\item{min_padj}{Minimum adjusted p value}
}
\value{
Character vector of gene names
}
\description{
Title
}
\examples{
\dontrun {
  getTopNGenes(results$Group_FAC_vs_Cont)
}

}
