% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoomByGroup.R
\name{getVoomByGroup}
\alias{getVoomByGroup}
\title{Run voomByGroup and prepare for ggplot2}
\usage{
getVoomByGroup(bulk, group, ...)
}
\arguments{
\item{bulk}{List object with counts in \code{bulk$dge$counts}, design in \code{bulk$md$design}, and samples in \code{bulk$dge$samples}.}

\item{group}{Name of column in bulk$dge$sample to group by (e.g. individual if running a repeated measures design)}

\item{...}{Arguments passed to voomByGroup.}
}
\value{
Long-format tibble with a columns for \code{group}, \code{x} and \code{y} of voom lines.
}
\description{
Run voomByGroup and reshape \code{voom.line} list output into a single tibble for input into \code{ggplot()}.
}
\note{
In practice, this is most useful as a diagnostic tool for pseudobulk data, where heteroscedasticity tends to be larger than for bulk data. See \code{\link[gencoreBulk:voomByGroup]{?voomByGroup}} for details.
}
\examples{
\dontrun{
vbg_data <- getVoomByGroupData(bulk, group = "SubjectID") \%>\% 
  plotVoomByGroupData()
}

}
