% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_qc_plots.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{Plot MDS as ggplot}
\usage{
ggplotMDS(
  dge,
  sampleID = "sampleID",
  gene.selection = "common",
  color,
  path,
  ...
)
}
\arguments{
\item{dge}{DGEList object}

\item{sampleID}{Label points by the value of this dge$samples column.}

\item{gene.selection}{\code{"common"} uses the same genes for all comparisons (i.e. PCA), \code{"pairwise"} chooses the top genes separately for each pairwise comparison between the samples (i.e MDS); see \code{\link[limma:plotMDS]{?limma::plotMDS()}} for details.}

\item{color}{Color points by the value of this dge$samples column.}

\item{path}{Connect points with the same value of this dge$samples column.}

\item{...}{Arguments passed to \code{\link[limma:plotMDS]{limma::plotMDS()}}}
}
\description{
Wrapper around \code{\link[limma:plotMDS]{limma::plotMDS()}} to enable ggplot2 formatting.
}
\examples{
\dontrun{
ggplotMDS(bulk$dge, gene.selection = "pairwise", 
          sampleID = "sampleID", color = "grp", path = "SubjectID")
}

}
