% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot.R
\name{gseaDotplot_single}
\alias{gseaDotplot_single}
\title{GSEA dotplot}
\usage{
gseaDotplot_single(
  result,
  filter_source = NULL,
  signif_only = FALSE,
  top_n = 20,
  min_size = 5,
  sig_cutoff = 0.05
)
}
\arguments{
\item{result}{A table with GSEA results from \code{fgsea::fgseaSimple()}}

\item{filter_source}{Character vector of pathway sources to filter for}

\item{signif_only}{If TRUE, only plot results with padj < \code{sig_cutoff}}

\item{top_n}{Show the top N results sorted by padj}

\item{min_size}{Minimum size of gene set to plot}

\item{sig_cutoff}{Threshold for significance, default to 0.05}
}
\value{
A ggplot object
}
\description{
Make dotplot from fGSEA result, showing top n pathways
}
\examples{
\dontrun{
res <- runfgsea(result, gmt.file)
gseaDotplot(res, filter_source = "HALLMARK")
}
}
