% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot.R
\name{gseaDotplot_single}
\alias{gseaDotplot_single}
\title{GSEA dotplot for a single contrast}
\usage{
gseaDotplot_single(
  result,
  filter_source = NULL,
  signif_only = FALSE,
  top_n = 20,
  min_size = 5,
  sig_cutoff = 0.05,
  use_shortened_pathway_names = FALSE,
  significance = c(0.05, 0.01, 0.001),
  breaks = c(0.1, 0.01, 0.001, 1e-04),
  cap_max = tail(breaks, 1)/10,
  p_val_col = "pval"
)
}
\arguments{
\item{result}{A table with GSEA results from \code{fgsea::fgseaSimple()}}

\item{filter_source}{Character vector of pathway sources to filter for}

\item{signif_only}{If TRUE, only plot results with p value < \code{sig_cutoff}}

\item{top_n}{Show the top N results sorted by pval}

\item{min_size}{Minimum size of gene set to plot}

\item{sig_cutoff}{Threshold for significance, default to 0.05}

\item{use_shortened_pathway_names}{Pull names from column 'pathway_short'
rather than pathway (if \code{runfgsea()} call had \code{breakdown_pathway_names} set
to \code{TRUE})}

\item{significance}{A vector of values to indicate significance with asterisks.
Each subsequent value will add an extra asterisk. E.g. \code{c(0.05, 0.01)} will
give one asteriks to values below 0.05 and two asterisks to values below 0.01.
If this is null, no asterisks will be plotted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:trans_new]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{p_val_col}{Column to use for significance values. Default 'pval'.}
}
\value{
A ggplot object
}
\description{
Make dotplot from an fGSEA result, showing top n pathways
}
\examples{
\dontrun{
res <- runfgsea(result, gmt.file)
gseaDotplot(res, filter_source = "HALLMARK")
}
}
