% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReadPerGeneFiles.R
\name{mappingBinsPlot}
\alias{mappingBinsPlot}
\title{Mapping bins plot
 
 Plot values of mapping bins for each sample. Takes the output of `parseReadPerGeneFiles()`}
\usage{
mappingBinsPlot(mapBins, title = "")
}
\description{
@param mapBins Data frame of 5 mapping bin categories for each sample
 @param title   Optional title for the plot
 
 @returns ggplot Plot
 @export
}
