% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCountsForHeatmap.R
\name{normalizeCountsForHeatmap}
\alias{normalizeCountsForHeatmap}
\title{Normalize an pre-processed expression matrix for heatmaps}
\usage{
normalizeCountsForHeatmap(
  count_matrix,
  metadata,
  group_var,
  baseline,
  remove_baseline = FALSE
)
}
\arguments{
\item{count_matrix}{Matrix of processed, log normalized counts data}

\item{metadata}{Metadata for samples in counts matrix}

\item{group_var}{Column of metadata to group samples by}

\item{baseline}{Level of \code{group_var} to use as baseline for normalization}

\item{remove_baseline}{Bool, remove baseline samples from output}
}
\value{
A numeric matrix
}
\description{
Normalize an RNA expression matrix by samples or groups of samples within the
matrix. The intended use case is to prepare data for plotting with heatmaps.
You define a grouping variable and a baseline level of that variable to normalize
other groups by. If there are more than one sample in the baseline group, the
median expression value is used to normalize other values.

This is NOT intended to normalized data as part of the upstream processing.
Data fed to this function should already be log normalized, as the baseline
normalization assumes the expression is on a log scale.
}
\examples{
\dontrun{
tmp <- normalizeCounts(assay(assays(obj.pbmc)$rld),
        obj.pbmc@colData,
        group_var = 'Timepoint', baseline = 'pre', 
        remove_baseline = TRUE)
heatmapFromGenelist(genes, tmp)
}
}
