% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCountsForHeatmap.R
\name{normalizeCountsForHeatmapByIndividual}
\alias{normalizeCountsForHeatmapByIndividual}
\title{Normalize an pre-processed expression matrix for heatmaps within individuals}
\usage{
normalizeCountsForHeatmapByIndividual(
  count_matrix,
  metadata,
  group_var,
  baseline,
  individual_var,
  remove_baseline = FALSE
)
}
\arguments{
\item{count_matrix}{Matrix of processed, log normalized counts data}

\item{metadata}{Metadata for samples in counts matrix}

\item{group_var}{Column of metadata to group samples by}

\item{baseline}{Level of \code{group_var} to use as baseline for normalization}

\item{individual_var}{Column of metadata to group individuals by}

\item{remove_baseline}{Bool, remove baseline samples from output}
}
\value{
A numeric matrix
}
\description{
Extension of \code{gencoreBulk::normalizeCounts} to normalize data within individual.
Samples are grouped by individual, then a baseline is defined on a per-individual
level. Samples for an individual are normalized by that individual's baseline.
If there are more than one sample in the individual's baseline, the median expression
value is used as the baseline (and you have a well designed study).
}
\details{
This is intended for use with time-course studies with repeated measures from
the same individual. Rather than normalizing by the group-level statistics,
you can normalize within individual to account for different baseline levels
of expression.

This is NOT intended to normalized data as part of the upstream processing.
Data fed to this function should already be log normalized, as the baseline
normalization assumes the expression is on a log scale.
}
\examples{
\dontrun{
tmp <- normalizeCountsByIndividual(assay(assays(obj.pbmc)$rld),
        obj.pbmc@colData,
        group_var = 'Timepoint', baseline = 'pre', 
        individual_var = 'Individual',
        remove_baseline = TRUE)
heatmapFromGenelist(genes, tmp)
}
}
