% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pcaPlotSimple}
\alias{pcaPlotSimple}
\title{Simple base for PCA plot}
\usage{
pcaPlotSimple(counts, metadata, xpc = 1, ypc = 2, ntop = 500)
}
\arguments{
\item{counts}{Expression data matrix. Colnames of counts should match rownames
of \code{metadata}}

\item{metadata}{Dataframe of metadata to associate with counts data. Rownames
of metadata should match colnames of \code{counts}.}

\item{xpc}{Numeric, which PC to plot on the x axis. Default 1}

\item{ypc}{Numeric, which PC to plot on the y axis. Default 2}

\item{ntop}{Number of genes to include in PCA. Default 500}
}
\value{
A ggplot object
}
\description{
Calculates PCA on counts data. Returns a ggplot object with the PCA data
and associated metadata. No additional geoms are included, this is meant
to be a blank template. You do need to provide metadata here so it can be
accessed with geoms you add to this output.
}
\examples{
\dontrun{
pcaPlot <- pcaPlotSimple(assay(assays(pbmc_subset)$vst), metadata = colData(pbmc_subset))

## Manually adding geoms
pcaPlot + geom_point(aes(color = Individual, shape=Timepoint)) 

}

}
