% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_qc_plots.R
\name{plotFilterByExpr}
\alias{plotFilterByExpr}
\title{Bar plot of filtered genes}
\usage{
plotFilterByExpr(dge, keep.exprs)
}
\arguments{
\item{dge}{DGEList object}

\item{keep.exprs}{Named logical vector with gene ids as names. Names must match rownames of \code{dge$counts}}
}
\description{
Generate a bar plot of genes passing or failing a filtering procedure, eg. \code{\link[edgeR:filterByExpr]{?edgeR::filterByExpr()}}.
}
\examples{
\dontrun{
keep.exprs <- filterByExpr(bulk$dge, design = bulk$md$design)
plotFilterByExpr(bulk$dge, keep.exprs)
}

}
