% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotPCAFromConfig}
\alias{plotPCAFromConfig}
\title{plot PCA from config}
\usage{
plotPCAFromConfig(analysis)
}
\arguments{
\item{analysis}{object with config options saved from the template format}
}
\value{
A ggplot
}
\description{
Returns plot of PCA using aesthetic labels specified in the analysis config
}
\examples{
\dontrun{
  plotPCAFromConfig(analysis)
}
 
}
