% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoomByGroup.R
\name{plotVoomByGroup}
\alias{plotVoomByGroup}
\title{Plot voomByGroup voom lines}
\usage{
plotVoomByGroup(vbg_data)
}
\arguments{
\item{vbg_data}{Long-format tibble with a columns for \code{group}, \code{x} and \code{y} of voom lines. In practice, this will be the output of \code{\link[gencoreBulk:getVoomByGroup]{?getVoomByGroup}}.}
}
\value{
\code{ggplot} plot with each group voomline plotting in a different color
}
\description{
Plot voomByGroup voom lines using \code{ggplot2} so that multiple plots can more flexibly be ggarranged.
}
\note{
In practice, this is most useful as a diagnostic tool for pseudobulk data, where heteroscedasticity tends to be larger than for bulk data. See \code{\link[gencoreBulk:voomByGroup]{?voomByGroup}} for details.
}
\examples{
\dontrun{
vbg_data <- getVoomByGroup(bulk, group = "SubjectID") \%>\% 
  plotVoomByGroup()
}

}
