% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read an excel workbook with sheets to a list of tables}
\usage{
read_excel_allsheets(filename, tibble = FALSE, ...)
}
\arguments{
\item{filename}{File in format .xls or .xlsx}

\item{tibble}{BOOL, Return the data as a tibble instead of a data.frame}

\item{\dots}{Additional arguments passed to \code{openxlsx::read.xlsx()}}
}
\value{
A list of data.frames (or tibbles)
}
\description{
Read an excel workbook with sheets to a list of tables
}
\examples{
\dontrun{
  dge_results <- read_excel_allsheets(here('outputs/dge.xlsx'))
}
}
