% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfgsea.R
\name{runfgsea}
\alias{runfgsea}
\title{runfgsea}
\usage{
runfgsea(
  result,
  pathways,
  minSize = 10,
  maxSize = 500,
  breakdown_pathway_names = FALSE,
  na_omit = TRUE
)
}
\arguments{
\item{result}{A DESeqResults result object}

\item{pathways}{List of gene sets to check.}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{breakdown_pathway_names}{Attempt to split pathway names into source and
pathway for easier reading}

\item{na_omit}{Bool, exclude genes with missing data in results? This checks
all values for NAs so be mindful of modified \code{result} objects}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway.
}
\description{
Perform GSEA on a DESeq2 result using the given pathways and fGSEA
The function takes about O(nk^{3/2}) time, where n is number of permutations
and k is a maximal size of the pathways. That means that setting 'maxSize'
parameter with a value of ~500 is strongly recommended.
}
\examples{
\dontrun{
runfgsea(results[[1]], gmt.file)
}

}
