% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoomByGroup.R
\name{voomByGroup}
\alias{voomByGroup}
\title{Modeling group heteroscedasticity in RNAseq data}
\usage{
voomByGroup(
  counts,
  group = NULL,
  design = NULL,
  lib.size = NULL,
  dynamic = NULL,
  normalize.method = "none",
  span = 0.5,
  save.plot = FALSE,
  print = TRUE,
  plot = c("none", "all", "separate", "combine"),
  col.lines = NULL,
  pos.legend = c("inside", "outside", "none"),
  fix.y.axis = FALSE,
  ...
)
}
\arguments{
\item{counts}{
    a numeric \code{matrix} containing raw counts, or an \code{ExpressionSet} containing raw counts, or a \code{DGEList} object.
    Counts must be non-negative and NAs are not permitted.}

\item{group}{vector or factor indicating groups to block voom estimates by.}

\item{design}{
    design matrix with rows corresponding to samples and columns to coefficients to be estimated.
    Defaults to \code{model.matrix(~0+counts$samples$group)} if \code{counts} is a DGEList, otherwise defaults to the unit vector meaning that all samples are treated as replicates.}

\item{lib.size}{
    numeric vector containing total library sizes for each sample.
    Defaults to the normalized (effective) library sizes in \code{counts} if \code{counts} is a \code{DGEList} or to the columnwise count totals if \code{counts} is a matrix.}

\item{dynamic}{determine groups dynamically (?)}

\item{normalize.method}{
    the microarray-style normalization method to be applied to the logCPM values (if any).
    Choices are as for the \code{method} argument of \code{normalizeBetweenArrays} when the data is single-channel.
    Any normalization factors found in \code{counts} will still be used even if \code{normalize.method="none"}.}

\item{span}{
    width of the smoothing window used for the lowess mean-variance trend.
    Expressed as a proportion between 0 and 1.}

\item{save.plot}{
    logical, should the coordinates and line of the plot be saved in the output?}

\item{print}{print name of each group as it's being processed.}

\item{plot}{plot each voom trend in a separate plot "separate", all in the same plot "combine", or plot none of them "none".}

\item{col.lines}{color of lines (?)}

\item{pos.legend}{legend position. Can be "inside", "outside", or "none".}

\item{fix.y.axis}{fix y axis to the same limits if plot = "separate".}

\item{...}{arguments passed to limma::voom.}
}
\description{
Accounting for heteroscedastic groups in RNAseq bulk and pseudobulk data.
}
\details{
This is an extension of \code{\link[limma:voom]{limma::voom()}} described in \code{\link[https://doi.org/10.1186/s13059-023-02949-2]{You et al. (2023)}}. The function source is a fork of the \code{\link[https://github.com/YOU-k/voomByGroup/blob/main/voomByGroup.R]{code from the supplemental materials}} from that paper, modified to output BCV values in the EList.
}
\examples{
\dontrun{
vbg <- voomByGroup(counts = bulk$dge$counts, 
}

}
