% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCountsTables.R
\name{writeCountTables}
\alias{writeCountTables}
\title{Write data to csvs}
\usage{
writeCountTables(
  raw_counts,
  analysis,
  outdir = "outputs",
  generate_GSEA_app_tables = FALSE,
  write_sample_table = TRUE
)
}
\description{
Takes the raw counts matrix and an analysis object with yaml config 
 options saved to write data to named file outputs.
 
 @param raw_counts matrix of raw count data
 @param analysis object with config options saved from the template format
 @param outdir folder to store outputs in, will be generated if it doesn't exist
 @param generate_GSEA_app_tables write files for compatibility with GSEA desktop app
 @param write_sample_table write sample metadata to a csv
}
\details{
@returns NULL

 @examples
 writeCountTables(raw_counts, analysis, generate_GSEA_app_tables = TRUE)
 
 @export
}
