% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDESeqResults.R
\name{writeDESeqResults}
\alias{writeDESeqResults}
\title{Write DESeqResults to excel sheet}
\usage{
writeDESeqResults(
  results,
  sheet_names = names(results),
  output_name = "DEG_results.xlsx",
  outdir = here::here("outputs"),
  padj_col = "padj",
  drop_NA = TRUE
)
}
\arguments{
\item{results}{List of DESeqResult objects}

\item{sheet_names}{List of names for sheets}

\item{output_name}{Name of file}

\item{outdir}{Location of file}

\item{padj_col}{Name of column with adjusted p values. For filtering and sorting}

\item{drop_NA}{Whether to exclude results with an adjusted p-value of \code{NA}}
}
\value{
\code{NULL}
}
\description{
Take a list of DESeqResults returned from \code{extractFilteredDESeqResults()}
and write them to an excel worksheet, with one sheet for each result.
}
\examples{
\dontrun{
 my_comparisons <- list('Group_FAC_vs_Cont',
 'Group_LiproxposFAC_vs_Cont',
 'Group_MTposFAC_vs_Cont')

 ## groups to include for greg's filter function, specific to each comparison
 ## should be levels of some group in colData(analysis$dds)
 filter_on_list <- list(c('FAC', 'Cont'),
                       c('LiproxposFAC', 'Cont'),
                       c('MTposFAC', 'Cont'))

 ## Get DESeq2 results for each comparison of interest
 ## intgroup specifies which metadata column your filter_on_list refers to
 results <- mapply(extractFilteredDESeqResults,
                   comparison = my_comparisons,
                   filt_groups = filter_on_list, 
                   intgroup = 'Group',
                   USE.NAMES = FALSE)
 names(results) <- lapply(results, function(x) x@metadata$contrast)
   writeDESeqResults(results)
}
}
