% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writefGSEAResults.R
\name{writefGSEAResults}
\alias{writefGSEAResults}
\title{Write fGSEA results to excel sheet}
\usage{
writefGSEAResults(
  results,
  sheet_names = names(results),
  output_name = "GSEA_results.xlsx",
  outdir = here::here("outputs")
)
}
\arguments{
\item{results}{List of fGSEA results objects}

\item{sheet_names}{List of names for sheets}

\item{output_name}{Name of file}

\item{outdir}{Location of file}
}
\value{
\code{NULL}
}
\description{
Take a list of fGSEA result tables returned from \code{runfgsea()}
and write them to an excel worksheet, with one sheet for each result.
}
\examples{
\dontrun{
 gsea_result_1 <- runfgsea(results$Group_FAC_vs_Cont,
  gmt.file,
  minSize = 1,
  breakdown_pathway_names = TRUE)
 
 gsea_result_2 <- runfgsea(results$Group_LiproxposFAC_vs_Cont,
  gmt.file,
  minSize = 1,
  breakdown_pathway_names = TRUE)

 gsea_results <- list('FAC vs Cont' = gsea_result_1, 'LiproxposFAC vs Cont' = gsea_result_2)
 writefGSEAResults(gsea_results)
}
}
