% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot_joint.R
\name{gseaDotplot_joint}
\alias{gseaDotplot_joint}
\title{GSEA dotplot}
\usage{
gseaDotplot_joint(
  gsea_results,
  pathway_order = NULL,
  x_order = NULL,
  significance = c(0.05, 0.01, 0.001),
  p_val_col = "pval"
)
}
\arguments{
\item{gsea_results}{A table with multiple GSEA results from \code{fgsea::fgseaSimple()}
combined by \code{combine_GSEA_results()}}

\item{pathway_order}{Order of pathways to plot}

\item{x_order}{Order of comparisons on X axis}

\item{significance}{A vector of values to indicate significance with asterisks.
Each subsequent value will add an extra asterisk. E.g. \code{c(0.05, 0.01)} will
give one asteriks to values below 0.05 and two asterisks to values below 0.01.
If this is null, no asterisks will be plotted.
@param p_val_col Column to use for significance values. Default 'pval'.}
}
\value{
A ggplot object
}
\description{
Make dotplot from fGSEA result, showing top n pathways
}
\examples{
\dontrun{
combine_GSEA_results <- function(gsea_results,
                                 pathways){
 gsea_results <- lapply(gsea_results, function(x){x \%>\% filter(pathway \%in\% pathways)})
 gsea_results <- data.table::rbindlist(gsea_results, idcol='ID')
}

pathways = c('REACTOME_SIGNALING_BY_GPCR', 'REACTOME_SIGNALING_BY_NTRKS')
gsea_results <- list('FAC vs Cont' = gsea_result, 'LiproxposFAC vs Cont' = gsea_result_2)
joint_GSEA_results <- combine_GSEA_results(gsea_results, pathways)
gseaDotplot_joint(joint_GSEA_results)
}
}
