% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot_joint.R
\name{gseaDotplot_joint}
\alias{gseaDotplot_joint}
\title{GSEA dotplot}
\usage{
gseaDotplot_joint(gsea_results)
}
\arguments{
\item{gsea_results}{A table with multiple GSEA results from \code{fgsea::fgseaSimple()}
combined by \code{combine_GSEA_results()}}
}
\value{
A ggplot object
}
\description{
Make dotplot from fGSEA result, showing top n pathways
}
\examples{
\dontrun{
combine_GSEA_results <- function(gsea_results,
                                 pathways){
 gsea_results <- lapply(gsea_results, function(x){x \%>\% filter(pathway \%in\% pathways)})
 gsea_results <- data.table::rbindlist(gsea_results, idcol='ID')
}

pathways = c('REACTOME_SIGNALING_BY_GPCR', 'REACTOME_SIGNALING_BY_NTRKS')
gsea_results <- list('FAC vs Cont' = gsea_result, 'LiproxposFAC vs Cont' = gsea_result_2)
joint_GSEA_results <- combine_GSEA_results(gsea_results, pathways)
gseaDotplot_joint(joint_GSEA_results)
}
}
