% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot.R
\name{gseaDotplot_joint}
\alias{gseaDotplot_joint}
\title{GSEA dotplot for multiple contrasts}
\usage{
gseaDotplot_joint(
  result,
  pathway_order = NULL,
  x_order = NULL,
  significance = c(0.05, 0.01, 0.001),
  breaks = c(0.1, 0.01, 0.001, 1e-04),
  cap_pvalues = TRUE,
  p_val_col = "pval",
  use_shortened_pathway_names = FALSE
)
}
\arguments{
\item{result}{A table with multiple GSEA results from \code{fgsea::fgseaSimple()}
combined by \code{combine_GSEA_results()}}

\item{pathway_order}{Order of pathways to plot}

\item{x_order}{Order of comparisons on X axis}

\item{significance}{A vector of values to indicate significance with asterisks.
Each subsequent value will add an extra asterisk. E.g. \code{c(0.05, 0.01)} will
give one asteriks to values below 0.05 and two asterisks to values below 0.01.
If this is null, no asterisks will be plotted.}

\item{breaks}{Values to mark for scaling radius
result,}

\item{cap_pvalues}{Bool; cap small pvalues to keep the range of dots smaller}

\item{p_val_col}{Column to use for significance values. Default 'pval'.}

\item{use_shortened_pathway_names}{Pull names from column 'pathway_short'
rather than pathway (if \code{runfgsea()} call had \code{breakdown_pathway_names} set
to \code{TRUE})}
}
\value{
A ggplot object
}
\description{
Make dotplot from fGSEA result, showing top n pathways
}
\examples{
\dontrun{
combine_GSEA_results <- function(gsea_results,
                                 pathways){
 gsea_results <- lapply(gsea_results, function(x){x \%>\% filter(pathway \%in\% pathways)})
 gsea_results <- data.table::rbindlist(gsea_results, idcol='ID')
}

pathways = c('REACTOME_SIGNALING_BY_GPCR', 'REACTOME_SIGNALING_BY_NTRKS')
gsea_results <- list('FAC vs Cont' = gsea_result, 'LiproxposFAC vs Cont' = gsea_result_2)
joint_GSEA_results <- combine_GSEA_results(gsea_results, pathways)
gseaDotplot_joint(joint_GSEA_results)
}
}
