% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReadPerGeneFiles.R
\name{mappingBinsPlot}
\alias{mappingBinsPlot}
\title{Mapping bins plot}
\usage{
mappingBinsPlot(mapBins, title = "")
}
\arguments{
\item{mapBins}{Data frame of 5 mapping bin categories for each sample}

\item{title}{Optional title for the plot}
}
\value{
ggplot object
}
\description{
Plot values of mapping bins for each sample. Takes the output of `parseReadPerGeneFiles()`
}
\examples{
mappingBinsPlot(analysis$mapBins)

}
