% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReadPerGeneFiles.R
\name{parseReadPerGeneFiles}
\alias{parseReadPerGeneFiles}
\title{Parse STAR readsPerGene.out files}
\usage{
parseReadPerGeneFiles(file.paths, library.type = "unstranded")
}
\arguments{
\item{file.paths}{Named vector of file paths to STAR output files, where names
correspond to sample names}

\item{library.type}{String of column to pull from in the output file. Options
should be 'unstranded', 'sense', or 'antisense'}
}
\value{
List of 2 data frames, one containing read mapping bins, the other
read per gene counts
}
\description{
Returns the counts per gene and read count bins for a list of STAR output files
}
\examples{
\dontrun{
   readfiles <- sapply(
     analysis$samplefileIDs,
     function(sid) {
       paste0(dir("my_STAR_output_dir",
                  pattern = sid, full.names = TRUE),
              "/", sid, 'readsPerGene.out.tab')})
              
   outs <- readCountFiles(readfiles, 'unstranded')
}

}
