% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneExpression.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Return a simple ggplot boxplot of gene expression}
\usage{
plotGeneExpression(
  gene,
  grouping,
  data = SummarizedExperiment::assays(analysis$dds)$rld,
  boxplot = TRUE,
  jitter = TRUE,
  axis_text_x = element_text(angle = 90, vjust = 0.5, hjust = 1)
)
}
\arguments{
\item{gene}{Gene to plot expression for}

\item{grouping}{Variable to group expression by. Must be a column of \code{data@colData}}

\item{data}{A DESeqDataSet object. Use this argument to select specific assays
to plot. The default is \code{assays(analysis$dds)$rld} (assumed to be the rlog transform).}

\item{boxplot}{Boolean, plot \code{geom_boxplot}?}

\item{jitter}{Boolean, plot \code{geom_jitter}?}

\item{axis_text_x}{Option to modify X axis text. Default rotates labels 90
degrees}
}
\value{
A ggplot object
}
\description{
Returns a simple ggplot of gene expression, optionally with jitter or
boxplots overlayed. These are optional in case you want to do other
aesthetics. This is meant to be minimal so you can adjust the aesthetics
as necessary, but it can still be used as a quick check.
}
\examples{
\dontrun{
## my object is called dds_full, so I have to manually provide data
plotGeneExpression('MAMU-A', grouping = 'groupIDs', data = assays(dds_full)$rld)
}

}
