% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotPCAFromConfig}
\alias{plotPCAFromConfig}
\title{plot PCA from config}
\usage{
plotPCAFromConfig(analysis, shapes = 0:25, alpha = 1, size = 5)
}
\arguments{
\item{analysis}{object with config options saved from the template format}

\item{shape_start}{Integer value to start range of shapes, default 0}
}
\value{
A ggplot
}
\description{
Returns plot of PCA using aesthetic labels specified in the analysis config
}
\examples{
\dontrun{
plotPCAFromConfig(analysis)
}

}
