% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plotPCAFromConfig}
\alias{plotPCAFromConfig}
\title{plot PCA from config}
\usage{
plotPCAFromConfig(
  analysis,
  shapes = 0:25,
  alpha = 1,
  size = 5,
  pc.1 = 1,
  pc.2 = 2
)
}
\arguments{
\item{analysis}{object with config options saved from the template format}

\item{shapes}{Shape values to pass to \code{ggplot2::scale_shape_manual()}}

\item{alpha}{Alpha to use for \code{ggplot2::geom_point()}}

\item{size}{Size to use for \code{ggplot2::geom_point()}}

\item{pc.1}{Principal component to plot on X axis}

\item{pc.2}{Principal component to plot on Y axis}
}
\value{
A ggplot
}
\description{
Returns plot of PCA using aesthetic labels specified in the analysis config
}
\examples{
\dontrun{
plotPCAFromConfig(analysis)
}

}
